% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_declustering.R
\name{declstr_random}
\alias{declstr_random}
\title{Random Declustering}
\usage{
declstr_random(spatial_object, cellsize, numpoints)
}
\arguments{
\item{spatial_object}{Object of class sfc or sf}

\item{cellsize}{Target cellsize in map units}

\item{numpoints}{Number of points that will be chosen from a grid cell}
}
\value{
Object of class sf
}
\description{
Function for random points selection from grid cells in a spatial object
}
