% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_declustering.R
\name{gridRandom}
\alias{gridRandom}
\title{Random Declustering}
\usage{
gridRandom(spatial_object, cellsize, numpoints)
}
\arguments{
\item{spatial_object}{Object of class sfc or sf}

\item{cellsize}{Definition of grid cell}

\item{numpoints}{Number of points that will be chosen from a grid cell}
}
\value{
Object of class sf
}
\description{
Function for random points selection from grid cells in a spatial object
}
