% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_declustering.R
\name{declstr_weighted}
\alias{declstr_weighted}
\title{Cell Declustering}
\usage{
declstr_weighted(spatial_object, cellsize)
}
\arguments{
\item{spatial_object}{Object of class sfc or sf}

\item{cellsize}{Target cellsize in map units}
}
\value{
Object of class sf
}
\description{
Function returns list of points with their weights and lengths from grid cell in spatial object
}
