% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_declustering.R
\name{gridWeighted}
\alias{gridWeighted}
\title{Cell Declustering}
\usage{
gridWeighted(spatial_object, cellsize)
}
\arguments{
\item{spatial_object}{Object of class sfc or path to object of class sfc}

\item{cellsize}{Definition of grid cell}
}
\value{
Object of class sfc
}
\description{
Function returns list of points with their weights and lengths from grid cell in spatial object
}
